Restructured testing env for mock API
