+++ src/setupTests.js
@@ -1,0 +1,11 @@
+// src/setupTests.js
+import server from './__mocks__/api/server';
+// Establish API mocking before all tests.
+beforeAll(() => server.listen());
+
+// Reset any request handlers that we may add during the tests,
+// so they don't affect other tests.
+afterEach(() => server.resetHandlers());
+
+// Clean up after the tests are finished.
+afterAll(() => server.close());
